Skies, dyes, and other load-time fixes

+++ World/MiscWorldgenRoutines.cs
@@ -480,7 +480,7 @@ namespace CalamityMod.World
         public static void UnderworldIslandHouse(int i, int j, int item)
         {
             ushort type = (ushort)ModContent.TileType<BrimstoneSlag>(); //tile
-            byte wall = (byte)ModContent.WallType<BrimstoneSlagWallUnsafe>(); //wall
+            int wall = ModContent.WallType<BrimstoneSlagWallUnsafe>(); //wall
             Vector2 vector = new Vector2((float)i, (float)j);
             int num = 1;
             if (WorldGen.genRand.Next(2) == 0)
